+++ src/main/java/com/uran/service/UserService.java
@@ -13,7 +13,7 @@ public interface UserService {
     void updateUser(User user);
     User getUser(int id);
     void deleteUser(int id);
-    Long count();
+    Integer count();
     List<User> getUsers(int page);
     List<User> getUsersBySearch(String userName);
 }
